+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Now the api key can be exposed in fetch instances if the EXPOSE_IN_FETCH_INSTANCES variable is set to true
 * Added option to generate qrcode as soon as the instance is created
 * The created instance token can now also be optionally defined manually in the creation endpoint
+* Route to send Sticker
 
 ### Fixed